+++ src/Language/Haskell/Names/LocalSymbolTable.hs
@@ -8,7 +8,6 @@ module Language.Haskell.Names.LocalSymbolTable
   ) where
 
 import qualified Data.Map as Map
-import Data.Monoid
 import Language.Haskell.Exts
 import Language.Haskell.Names.SyntaxUtils (dropAnn)
 import Language.Haskell.Names.Types